backend/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
groq==0.31.0
python-dotenv==1.1.1
